+++ library/core/src/main/java/com/google/android/exoplayer2/ExoPlayerImpl.java
@@ -724,7 +724,7 @@ import java.util.concurrent.CopyOnWriteArraySet;
   }
 
   @Override
-  public int getMaxSeekToPreviousPosition() {
+  public long getMaxSeekToPreviousPosition() {
     return C.DEFAULT_MAX_SEEK_TO_PREVIOUS_POSITION_MS;
   }
